requirements.txt
Django~=3.0.8

django-debug-toolbar~=2.2